+++ java-manta-client/src/main/java/com/joyent/manta/config/MapConfigContext.java
@@ -73,6 +73,16 @@ public class MapConfigContext implements ConfigContext {
      */
     public static final String MANTA_HTTPS_CIPHERS_KEY = "https.cipherSuites";
 
+    /**
+     * Property key for disabling HTTP signatures.
+     */
+    public static final String MANTA_NO_AUTH_KEY = "manta.no_auth";
+
+    /**
+     * Property key for disabling native code support for generating signatures.
+     */
+    public static final String MANTA_NO_NATIVE_SIGS_KEY = "http.signature.native.rsa";
+
     // I know manually adding them all sucks, but it is the simplest operation
     // for a shared library. We could do all sorts of complicated reflection
     // or annotation processing, but they are error-prone.
@@ -84,7 +94,8 @@ public class MapConfigContext implements ConfigContext {
             MANTA_KEY_PATH_KEY, MANTA_TIMEOUT_KEY, MANTA_RETRIES_KEY,
             MANTA_MAX_CONNS_KEY, MANTA_PRIVATE_KEY_CONTENT_KEY,
             MANTA_PASSWORD_KEY, MANTA_HTTP_TRANSPORT_KEY,
-            MANTA_HTTPS_PROTOCOLS_ENV_KEY, MANTA_HTTPS_CIPHERS_KEY
+            MANTA_HTTPS_PROTOCOLS_ENV_KEY, MANTA_HTTPS_CIPHERS_KEY,
+            MANTA_NO_AUTH_KEY, MANTA_NO_NATIVE_SIGS_KEY
     };
 
     /**
@@ -129,7 +140,7 @@ public class MapConfigContext implements ConfigContext {
     @Override
     public String getPrivateKeyContent() {
         return normalizeEmptyAndNullAndDefaultToStringValue(
-                MANTA_PRIVATE_KEY_CONTENT_KEY, MANTA_PRIVATE_ENV_KEY_CONTENT);
+                MANTA_PRIVATE_KEY_CONTENT_KEY, MANTA_PRIVATE_KEY_CONTENT_ENV_KEY);
     }
 
     @Override
@@ -192,6 +203,28 @@ public class MapConfigContext implements ConfigContext {
     public String getHttpsCipherSuites() {
         return normalizeEmptyAndNullAndDefaultToStringValue(
                 MANTA_HTTPS_CIPHERS_KEY, MANTA_HTTPS_CIPHERS_ENV_KEY);
+    }
+
+    @Override
+    public Boolean noAuth() {
+        Boolean mapValue = MantaUtils.parseBooleanOrNull(backingMap.get(MANTA_NO_AUTH_KEY));
+
+        if (mapValue != null) {
+            return mapValue;
+        }
+
+        return MantaUtils.parseBooleanOrNull(backingMap.get(MANTA_NO_AUTH_ENV_KEY));
+    }
+
+    @Override
+    public Boolean disableNativeSignatures() {
+        Boolean mapValue = MantaUtils.parseBooleanOrNull(backingMap.get(MANTA_NO_NATIVE_SIGS_KEY));
+
+        if (mapValue != null) {
+            return mapValue;
+        }
+
+        return MantaUtils.parseBooleanOrNull(backingMap.get(MANTA_NO_NATIVE_SIGS_ENV_KEY));
     }
 
     /**